Rename DynamicPartitionPruningTest to DynamicPartitionPruningCoverage
This avoids Test[...]Test names.

I used the following script.

```
set -euo pipefail

files=()
while IFS= read -r f; do
  files+=("$f")
done < <(git ls-files '*DynamicPartitionPruningTest.java')

names=()
for f in "${files[@]}"; do
  base=$(basename "$f" .java)
  names+=("$base")
  new="${f%DynamicPartitionPruningTest.java}"
  new="${new}DynamicPartitionPruningCoverage.java"
  git mv "$f" "$new"
done

for n in "${names[@]}"; do
  new="${n%DynamicPartitionPruningTest}DynamicPartitionPruningCoverage"
  git grep -lz "$n" | while IFS= read -r -d '' file; do
    perl -i -pe "s/\b\Q${n}\E\b/${new}/g" "$file"
  done
done
```
